Backend/Services/ExtraProcess/requirements.txt
numpy
protobuf
grpcio
requests
certifi
opencv-python
opencv-contrib-python
pytz
vidgear